Move alla hash utils in a class called Hasher
