+++ lib/screens/providers/hasher_provider.dart
@@ -1,0 +1,4 @@
+import 'package:flut_crack/data/hasher.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+
+final hasherProvider = Provider<Hasher>((ref) => Hasher());
